Fix `.classList` errors in old Android and IE10
Android doesn’t support `element.classList` at all, IE10 does but not
for SVG elements. jQuery does not support `.addClass()` for SVG
elements, but using the attribute method to set the class of the SVG
element works. This also means we don’t have to check that an element
in the data exists in the SVG, since it’s a jQuery object and won’t
return undefined.